Extract inference corpus preflight pipeline from run.py.
Move GT extraction, coverage, and Tier 2 checks into pipeline_preflight.py
so run.py stays an orchestrator, and initialize coverage_block before gated
branches to avoid fragile scoping.
